Enable support for lowmem in conf files
Once command line arguments are restored to core.runtime, it's possible
to use D logic to re-execute the executable if -lowmem is detected and
the GC is present.  This should only be impactful to those who need it
in a conf file, and introduces no overhead, aside from extra memory
allocation, when it's not.